requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.35.0
accelerate==0.24.0